Implemented Validators and Exceptions in Services

+++ src/main/java/service/ClientServiceDAOImp.java
@@ -3,6 +3,8 @@ package service;
 import models.Client;
 import models.Reservation;
 import repository.ClientDAO;
+import service.Exceptions.ClientException;
+import service.Validators.ClientValidator;
 
 import java.time.LocalDateTime;
 import java.util.List;
@@ -16,27 +18,44 @@ public class ClientServiceDAOImp implements ClientServiceDao {
     }
 
     @Override
-    public Client add(Client client) {
+    public Client add(Client client) throws ClientException {
+        if(!ClientValidator.IsClientValid(client)){
+            throw new ClientException("Client data is invalid");
+        }
         return clientServiceDao.add(client);
     }
 
     @Override
-    public Client update(Client client) {
+    public Client update(Client client)throws ClientException {
+        if(!ClientValidator.IsClientValid(client)){
+            throw new ClientException("Client data is invalid");
+        }
+
         return clientServiceDao.add(client);
     }
 
     @Override
-    public List<Reservation> reservationByClient(ClientDAO clientDAO) {
+    public List<Reservation> reservationByClient(Client client) throws ClientException {
+        if(!ClientValidator.IsClientValid(client)){
+            throw new ClientException("Client data is invalid");
+        }
         return null;
     }
 
     @Override
-    public Client delete(Client client) {
+    public Client delete(Client client)throws ClientException {
+        if(!ClientValidator.IsClientValid(client)){
+            throw new ClientException("Client data is invalid");
+        }
+
         return clientServiceDao.delete(client);
     }
 
     @Override
-    public List<Reservation> reservationByClientInSpecificPeriod(Client client, LocalDateTime startDate, LocalDateTime endDate) {
+    public List<Reservation> reservationByClientInSpecificPeriod(Client client, LocalDateTime startDate, LocalDateTime endDate) throws ClientException{
+        if(!ClientValidator.IsClientValid(client)){
+            throw new ClientException("Client data is invalid");
+        }
         return null;
     }
 }
